fix(proj): add new user info, update db
Lots of changes have been made. The database has been changed from mongo to mssql. The was that data is stored has been updated to work appropriately, and nodemail has been added for passwrod reset functionality. Promark user info tracking has been added that interacts properly with new tables in the database. Mongoose will be entirely removed but redux will stay in place for the use of store and persist. New user controller were added including promark user controllers and password reset controllers. Roles have been adjusted to match the database.

BREAKING CHANGE: As a migration to a new database occurred, functionality may need to be updated accross the board to accomodate the change to the new database. The way authentication AND authorization is handled may need to be updated in their respective controllers as well. The env file has changed and must be update on all systems.