refactor(proj): migrate to vite
create-react-app has been depracated as of Feb 14, 2025. The project has been migrate from CRA to Vite. All files have been changed and currently the app is working as expected.

BREAKING CHANGE: A lot of react components will fail to work without import React from 'react'; included in the top of the file. This was not necessary previously. Currently irt has been added on top of EVERY file until it is determined that they can be removed.