feat(auth-persist): add auth persist
Auth persist has been added by using a "keep me logged in" checkbox. Without the box being checked, you are automatically logged out in 15 minutes when reauthentication is attempted. If the box is checked, the authentication is allowed.

BREAKING CHANGE: Cookies in the backend must be updated to handle login persist.