feat(auth-persist): cookie persist enabled
Cookie persist has been enabled so that a "keep me logged in" box must be checked in orer to stay logged in. The refresh token timeout has been changed to 12 hours.

BREAKING CHANGE: front end changes have been committed, be sure to pull the changes in order to allow the cookie to be read by the server