feat(live-projects-ui/api): add filter, change queries/handlers
The LiveProjects.jsx has been split and a components directory was added for current and future component implementations. The filter has been updated to allow all and currently selected projects to be queried, this will allow a smoother look and feel to the filter. The controller has been updated to use a helper function to create the handler. This reduces the amount of code needed in the file. new queries were made to get all and filtered data in the models. Routes have been adjusted accordingly.