fix(ui/tables/proj+summary-rep): fix layout + tables
The focus of this update is for mobile viewing. The original layout was too clunky and crowded. The new look and feel has tables scaled for their individual pages. This allows the tables to take up the proper width and have enough padding to seem less overcrowded. Text alignemtn has been centered instead of left justified. Project name now has a size limit of 8 character + ... (11 total) to prevent overcrowding on the summary report.