fix(ui/api-auth): security update
Updated token storage and date tracking. Previously your access token was not stored in the database, and while it isnt required to be stored it allows a nice way to see who is currently logged in without having a logged in column. The refresh token and access token are not set to clear when the user logs out manually, or is logged out automatically. The dateUpdated column will also be set tot he current date upon the action being taken.