fix(ui/navbar): fix nav
Fixed a bug where the navbar wouldnt navigate when clicking links. This was caused by the button drawing over the links (higher z-index) and preventing a click. The navbare has been switched to an <aside> for semantics and the compenent has been given a clickable area that shrinks when open. The styling has been updated as well.