feat(proj/reports): add summary store, add dbconn features, add model
Added a store for gpcph, dates, and live toggles. This will be used to routing. A model has been added for production reports, and routing has been set up for it as well. dbconn retry features would occasionally block requests, parameters were added for more fine tuned control. A bug where the abbreviatedDate was 1 behind was fixed. Fetch data was simplified in the logic files.