fix(backend/quota): fix calc for web only
Fixed calculations for web only projects. Since a com file is not used some adjustments needed to be made to account for it. If a com file has a length of 0 then a webCom is used instead. STYPE MUST still be declared in web projects, typically STYPE=3 with a label that has the objective in parentheses () e.g. (1000). The rule where 1000 is subtracted from objectives is ignored on web only projects.