fix(server/quotas): fix userid, label regex, add unused, set visibleStructure
Fixed userid to be used to determing what projects are visible to external users. These are linked as project by user. Empty structure has been replaced with visible structure. By default structure visibility is set to false, but if it exists then the visibility is set to true. Unused was added to the data to track unused saample. This request was specifically from Sarah. The regex for labels has been updated. It can filter for all types such as (Min #) (Max#) (T:#) (#) and doesnt care about white space.