fix(ui/quotas): moved dir, add unused, fix toggle, fix format
Files were moved to a different directory. This is to keep quota related code in the quota management folder, but to separate viewing and publishing. Publishing will be an internal only portion. Toggles were fixed so that the All Columns toggle will now work as intended. Some default columns were removed from external users. Styling updated were made to the table. The projectlist has been changed to use a lazy loader with userid for external users