fix(server/quotas): fix project list bug
Fixed a bug where project list wasn't properly fetched. This was due to the userId being undefined as a string rather than a type of undefined. A clean query helper function was made and added to utils to be used in other fiels where setting undefined from a string may be necessary.