feat(proj/quotas): move skip row logic, fix api call
Moved the skip logic to the server rather than the browser. Previously the front end would hide rows that werent needed, however some data storage is needed to be calculated based off of hidden criteria. This task has been offloaded to the server rather than the client. The data is now being removed so it wont be available for the front end to view as it will no longer be in the response. This will allow future fine grained controls over which rows are visible. The api call has been updated to send the isInternalUser state flag. This is a simply true or false value. Anyone who has a role of external, regardless of other roles, will not be considered an internal user.