fix(proj): upd list fetching
Project list fetching has been moved into a project info section. This is its own component with its own routing and controllers. This was done because the portal will be heavily relient on user specific acces and neesd to remain consistent. This will also reduce the use of redundant code. The pie chart has been updated with some new styling info for a better look and feel.