fix(proj/roles): upd roles config
Roles were previously stored in a file. This has been completely updated to use the data from the database instead. This functionally was supposed to be working already, but was recently discovered to be missing. The proper updates have been pushed to ensure that this is consistent accross the whole portal.