fix(CI/CD): fix build
The build steps were previously all jumbled and would not properly build the test environment. This has been remedied and will now work appropriately.

BREAKING CHANGE: The main.yml is going to be tested in this push. It is not confirmed to work properly however i have confirmed that all my configuration files will spin up the server correctly when done manually.