fix(sample-auto): fix header to allow non match
A non matching header override was added to allow files to be uploaded even if they dont all contain the same headers. This was done because there is no guarantee that both files will have the same data, but they both still needs ot be uploaded where they do match.