fix(sample): upd sample auto
Major changes were made to the sample automation process. A lot of bug fixes were implements. Batches have been fixed to represent properly stratified data. Tarrance has been updated to have a more streamlined process that doesn't include an age threshold or source requirement for LL and CELL splits. The $N column will also use the proper WPHONE column for calculation. File id base functionality has been added but is not yet in final form. It will currently auto increment. Page styling has seen significant updates to make the user experience better. There is less clutter and more room to work with. The sample configuration header selection for extraction has been made to highlight the heaers that are flagged for export. This is for future updates that will be used to auto select based on our standardized criteria. This will make it easier to see what is and isnt selected. A process was updated to split land and cell into their own tables rather than try to work with a single table in sql. This was done to make stratification easier and make a simpler process for exportation and data viewing. This may be reverted in the future. A modal was missed in the last commit, and whilst i can just amend it, i have already typed this long message in the console and thus it will be pushed in this commit. Various styling updates have occured to make the entire page more user friendly.