fix(summary-reports): fix flicker and load bug, upd style
The previous version would check for invalid data if the received rows were not the same. 5 requests would be sent in sequence to verify the data was accurate upon updating. This would cause flickering that was tough to look at. This has been changed to be handled more gracefully. The styling has also been updated toa more modern look and feel.