fix(callid): fix assignment bug
Call ids were previoiusly not showing in the assignments tab, but if a callid existed in that slot then a remove button would show. The remove button has been removed as it should not exist. This is because we only want to change call ids and never remove call ids entirely. The bug that wouldnt display the callid is also fixed. Proper callids display in the proper slot on the editing modal. The inventory table now shows a status column where in use will be displayed rather than an in use column. This was sdone to prevent redundant database calls as the data was already included in the status colmn anyway. Upon assignment the status will be updated to be "in use" with a code of 2.

BREAKING CHANGE: An issue has arrised where the callid status needs to be changed after shift. In this case, for simplest use we will run a procedure at 6am the every day to verify if a project is past its end date, and if so, tag all in use numbers back to available.